requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
albumentations=0.5.2=pypi_0
numpy=1.19.2=py38h54aff64_0
opencv-python=4.5.1.48=pypi_0
pandas=1.2.2=py38ha9443f7_0
pip=21.0.1=py38h06a4308_0
python=3.8.8=hdb3f193_4
pytorch=1.7.1=py3.8_cuda11.0.221_cudnn8.0.5_0
scikit-learn=0.23.2=py38h0573a6f_0
torchsummary=1.5.1=pypi_0
tqdm=4.56.0=pyhd3eb1b0_0